+++ src/snowflake/snowpark/modin/plugin/__init__.py
@@ -5,6 +5,7 @@
 import inspect
 import sys
 from typing import Union, Callable, Any
+import warnings
 
 from packaging import version
 
@@ -183,8 +184,19 @@ from modin.core.storage_formats.pandas.query_compiler_caster import (  # isort: 
 )
 from modin.config import AutoSwitchBackend  # isort: skip  # noqa: E402
 
+HYBRID_WARNING = (
+    "Snowpark pandas now runs with hybrid execution enabled by default, and will perform certain operations "
+    + "on smaller data using local, in-memory pandas. To disable this behavior and force all computations to occur in "
+    + "Snowflake, run this line:\nfrom modin.config import AutoSwitchBackend; AutoSwitchBackend.disable()"
+)
+
+warnings.filterwarnings("once", message=HYBRID_WARNING)
+
 if AutoSwitchBackend.get_value_source() is ValueSource.DEFAULT:
-    AutoSwitchBackend.disable()
+    AutoSwitchBackend.enable()
+
+if AutoSwitchBackend.get():
+    warnings.warn(HYBRID_WARNING, stacklevel=1)
 
 # Hybrid Mode Registration
 # In hybrid execution mode, the client will automatically switch backends when a